requirements.txt
Pillow
google-generativeai